+++ CHANGELOG.md
@@ -3,6 +3,9 @@ CHANGELOG for Sulu Pricing Bundle
 
 * dev-develop
 
+    * FEATURE Adopted GroupedItemsPriceCalculator: Calculates now all prices
+              (totalPrice, totalNetPrice, totalRecurringPrice, totalRecurringNetPrice, shippingCosts, taxes) 
+    * FEATURE Refactored PriceCalculationManager: Moved calculation to ItemPriceCalculator
     * FEATURE Moved manager to own folder
 
 * 0.4.2 (2016-07-28)